+++ src/10-Core/Wtq/Services/WtqOptionsSaveService.cs
@@ -1,3 +1,5 @@
+#pragma warning disable
+
 using System.Collections;
 using System.Collections.Concurrent;
 using System.Reflection;
