+++ src/10-Core/Wtq/WtqConstants.cs
@@ -55,7 +55,7 @@ public static class WtqConstants
 			Description = "Windows Terminal",
 			Image = "/example_apps/windows_terminal.jpg",
 			Link = new("https://learn.microsoft.com/en-us/windows/terminal/install"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Windows Terminal",
@@ -72,7 +72,7 @@ public static class WtqConstants
 			Description = "The original PowerShell, (Windows-only).",
 			Image = "/example_apps/powershell_5.jpg",
 			Link = new("https://learn.microsoft.com/en-us/powershell/scripting/install/installing-powershell-on-windows?view=powershell-5.1"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "PowerShell 5",
@@ -88,7 +88,7 @@ public static class WtqConstants
 			Description = "The cross-platform PowerShell, (version 6 and up, a.k.a. 'PowerShell Core').",
 			Image = "/example_apps/powershell_core.png",
 			Link = new("https://learn.microsoft.com/en-us/powershell/scripting/install/installing-powershell-on-windows?view=powershell-7.4"),
-			Os = [ OSPlatform.Windows, ], // On Linux, terminal emulator and shell are more separate than on Windows.
+			Os = [OSPlatform.Windows,], // On Linux, terminal emulator and shell are more separate than on Windows.
 			Factory = () => new()
 			{
 				Name = "PowerShell Core",
@@ -104,12 +104,12 @@ public static class WtqConstants
 			Description = "Cross-platform GPU-accelerated terminal emulator.",
 			Image = "/example_apps/wezterm.png",
 			Link = new("https://wezfurlong.org/wezterm/index.html"),
-			Os = [ OSPlatform.Windows, ], // On Linux, terminal emulator and shell are more separate than on Windows.
+			Os = [OSPlatform.Windows,], // On Linux, terminal emulator and shell are more separate than on Windows.
 			Factory = () => new()
 			{
 				Name = "Wezterm",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -120,12 +120,12 @@ public static class WtqConstants
 			Description = "Cross-platform, fast, feature-rich, GPU based terminal.",
 			Image = "/example_apps/kitty.png",
 			Link = new("https://sw.kovidgoyal.net/kitty/"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Wezterm",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -136,12 +136,12 @@ public static class WtqConstants
 			Description = "Ptyxis is a terminal for GNOME that focuses on ease-of-use in a world of containers.",
 			Image = "/example_apps/ptyxis.webp",
 			Link = null,
-			Os = [ OSPlatform.Linux, ],
+			Os = [OSPlatform.Linux,],
 			Factory = () => new()
 			{
 				Name = "Ptyxis",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -152,12 +152,12 @@ public static class WtqConstants
 			Description = "The default console emulator for KDE.",
 			Image = "/example_apps/konsole.png",
 			Link = null,
-			Os = [ OSPlatform.Linux, ],
+			Os = [OSPlatform.Linux,],
 			Factory = () => new()
 			{
 				Name = "Konsole",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -169,7 +169,7 @@ public static class WtqConstants
 			Title = "Windows Explorer",
 			Description = "Windows explorer, the default file browser on Windows.",
 			Image = "/example_apps/windows_explorer.png",
-			Os = [ OSPlatform.Windows, ], // On Linux, terminal emulator and shell are more separate than on Windows.
+			Os = [OSPlatform.Windows,], // On Linux, terminal emulator and shell are more separate than on Windows.
 			Factory = () => new()
 			{
 				Name = "Windows Explorer",
@@ -185,12 +185,12 @@ public static class WtqConstants
 			Description = "Midnight Commander-inspired file manager.",
 			Image = "/example_apps/double_commander.png",
 			Link = new("https://doublecommander.com"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Double Commander",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -201,12 +201,12 @@ public static class WtqConstants
 			Description = "File manager with support for multiple panes, featuring lots of customization.",
 			Image = "/example_apps/q_dir.png",
 			Link = new("http://q-dir.com/"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Double Commander",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -217,12 +217,12 @@ public static class WtqConstants
 			Description = "The default file manager for KDE.",
 			Image = "/example_apps/dolphin.jpg",
 			Link = null,
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Dolphin",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -235,12 +235,12 @@ public static class WtqConstants
 			Description = "Music streaming service.",
 			Image = "/example_apps/spotify.jpg",
 			Link = new("https://www.spotify.com/us/download/"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Spotify",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -251,12 +251,12 @@ public static class WtqConstants
 			Description = "Chat service.",
 			Image = "/example_apps/discord.png",
 			Link = new("https://discord.com/"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Discord",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -267,12 +267,12 @@ public static class WtqConstants
 			Description = "Password manager.",
 			Image = "/example_apps/keepassxc.png",
 			Link = new("https://keepassxc.org/"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "KeePassXC",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -283,12 +283,12 @@ public static class WtqConstants
 			Description = "Task manager on steroids.",
 			Image = "/example_apps/process_explorer.jpg",
 			Link = new("https://learn.microsoft.com/en-us/sysinternals/downloads/process-explorer"),
-			Os = [ OSPlatform.Windows, ],
+			Os = [OSPlatform.Windows,],
 			Factory = () => new()
 			{
 				Name = "Process Explorer",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 
@@ -299,21 +299,21 @@ public static class WtqConstants
 			Description = "Default task manager for KDE.",
 			Image = "/example_apps/system_monitor.jpg",
 			Link = null,
-			Os = [ OSPlatform.Linux, ],
+			Os = [OSPlatform.Linux,],
 			Factory = () => new()
 			{
 				Name = "System Monitor",
 				Hotkeys = [new() { Modifiers = KeyModifiers.Control, Key = Keys.D1, }],
-				FileName = "",
+				FileName = "TODO",
 			},
 		},
 	];
 
 	private static readonly Random _r = new();
 
-	public static string GetRandomPlaceholderName() => PlaceholderNames[_r.Next(0, PlaceholderNames.Length)];
+	public static string GetRandomPlaceholderName() => _placeholderNames[_r.Next(0, _placeholderNames.Length)];
 
-	private static readonly string[] PlaceholderNames =
+	private static readonly string[] _placeholderNames =
 	[
 		"Akira",
 		"Ambassador",
@@ -393,21 +393,4 @@ public static class WtqConstants
 		"Zodiac",
 	];
 
-}
-
-public class WtqAppExample
-{
-	public required string Title { get; init; }
-
-	public required string Description { get; init; }
-
-	public required Func<WtqAppOptions> Factory { get; init; }
-
-	public required ICollection<OSPlatform> Os { get; init; }
-
-	public Uri? Link { get; set; }
-
-	public string? Image { get; set; }
-
-	public string? Icon { get; set; }
 }
