+++ src/10-Core/Wtq/WtqPaths.cs
@@ -36,7 +36,7 @@ public static class WtqPaths
 	/// <summary>
 	/// Returns path to the WTQ config file, as specified by an environment variable.
 	/// </summary>
-	public static string WtqConfigFromEnvVar => WtqEnv.ConfigFile;
+	public static string WtqConfigFromEnvVar => WtqEnv.ConfigFileFromEnv;
 
 	/// <summary>
 	/// Ordered list of possible paths to WTQ configuration file.<br/>
