+++ src/20-Services/Wtq.Services.Win32/Native/User32.cs
@@ -58,7 +58,7 @@ public static class User32
 	[return: MarshalAs(UnmanagedType.Bool)]
 	public static extern bool SetWindowPos(nint hWnd, nint hWndInsertAfter, int x, int y, int cx, int cy, uint uFlags);
 
-	[DllImport("user32.dll", SetLastError = true)]
+	[DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
 	public static extern int SetWindowText(IntPtr hWnd, string text);
 
 	[DllImport("user32.dll", SetLastError = true)]
