+++ src/20-Services/Wtq.Services.WinForms/TrayIcon.cs
@@ -21,10 +21,9 @@ public sealed class TrayIcon : IDisposable
 			var contextMenu = new ContextMenuStrip();
 
 			contextMenu.Items.AddRange(
-			[
 				CreateItem(
 					$"Version {WtqConstants.AppVersion}",
-					() => {},
+					() => { },
 					enabled: false),
 
 				new ToolStripSeparator(),
@@ -57,8 +56,7 @@ public sealed class TrayIcon : IDisposable
 
 				CreateItem(
 					"Exit",
-					() => lifetime.StopApplication()),
-			]);
+					() => lifetime.StopApplication()));
 
 			// Tray Icon
 			_notificationIcon = new NotifyIcon()
